moved sh_popen and sh_system to their own module in dttools called sh_popen
