+++ batch_job/src/batch_job_amazon_batch.c
@@ -16,6 +16,7 @@ See the file COPYING for details.
 #include "itable.h"
 #include "hash_table.h"
 #include "fast_popen.h"
+#include "sh_popen.h"
 
 
 #include <time.h>
@@ -39,10 +40,6 @@ static int instID;
 static char* bucket_name = NULL;
 static struct hash_table* submitted_files = NULL;
 static int HAS_SUBMITTED_VALUE = 1;
-
-static FILE *sh_popen(const char *command);
-static int sh_pclose(FILE * file);
-static int sh_system(const char* command);
 
 
 union amazon_batch_ccl_guid{
@@ -565,120 +562,3 @@ const struct batch_queue_module batch_queue_amazon_batch = {
 		batch_fs_amazon_batch_unlink,
 	},
 };
-
-static int sh_system(const char* command) {
-	int pid;
-	int res = 0;
-	pid = fork();
-	if (pid == 0) {//child
-		char* argv[4];
-		argv[0] = "sh";
-		argv[1] = "-c";
-		argv[2] = (char*)command;
-		argv[3] = NULL;
-
-		res = execvp(argv[0], argv);
-		if (res < 0) {
-			perror("Batch Job SH_system past execvp: ");
-		}
-	} else if (pid > 0) {//parent
-		struct process_info* pres;
-		while((pres = process_waitpid(pid, 0)) == 0);
-		if (WIFEXITED(pres->status)) {
-			return WEXITSTATUS(pres->status);
-		}
-	} else {
-		fprintf(stderr, "error in forking\n");
-	}
-	return -17;
-
-}
-
-static struct itable *process_table = 0;
-
-static FILE *sh_popen(const char *command)
-{
-	pid_t pid;
-	int argc;
-	char **argv;
-	int fds[2];
-	char cmd[4096];
-	int result;
-
-	strcpy(cmd, command);
-
-	if(string_split_quotes(cmd, &argc, &argv) < 1)
-		return 0;
-	
-	argv = malloc(sizeof(char*)*4);
-	argv[0]="sh";
-	argv[1]="-c";
-	argv[2] = (char*)command;
-	argv[3] = NULL;
-	argc = 4;
-
-	if(argc < 1)
-		return 0;
-
-	result = pipe(fds);
-	if(result < 0) {
-		free(argv);
-		return 0;
-	}
-
-	pid = fork();
-	if(pid > 0) {
-		free(argv);
-		close(fds[1]);
-
-		if(!process_table)
-			process_table = itable_create(0);
-
-		itable_insert(process_table, fds[0], (void *) (PTRINT_T) pid);
-		return fdopen(fds[0], "r");
-
-	} else if(pid == 0) {
-
-		int i;
-
-		close(0);
-		i = dup2(fds[1], STDOUT_FILENO); if(i < 0) perror("Error Dup2 stdout");
-		i = dup2(fds[1], STDERR_FILENO); if(i < 0) perror("Error Dup2 stderr");
-		close(fds[1]);
-		close(fds[0]);
-
-		i = execvp(argv[0], argv);
-		if (i<0) perror("Error in execvp");
-		_exit(1);
-
-	} else {
-		free(argv);
-		return 0;
-	}
-}
-
-static int sh_pclose(FILE * file)
-{
-	pid_t pid;
-	struct process_info* result;
-
-	pid = (PTRINT_T) itable_remove(process_table, fileno(file));
-
-	fclose(file);
-
-	while(1) {
-		while((result = process_waitpid(pid,0)) == 0);
-		if(WIFEXITED(result->status)) {
-			return WEXITSTATUS(result->status);
-		} else {
-			if(errno == EINTR) {
-				continue;
-			} else {
-				break;
-			}
-		}
-	}
-
-	errno = ECHILD;
-	return -1;
-}
