+++ dttools/src/Makefile
@@ -93,6 +93,7 @@ SOURCES = \
 	set.c \
 	sha1.c \
 	shell.c \
+	sh_popen.c\
 	sigdef.c \
 	sleeptools.c \
 	sort_dir.c \
